[AWS|DynamoDB] ConsumedCapacityUnits should be a Float

+++ tests/aws/requests/dynamodb/item_tests.rb
@@ -61,7 +61,7 @@ Shindo.tests('Fog::AWS[:dynamodb] | item requests', ['aws']) do
     end
 
     @query_format = {
-      'ConsumedCapacityUnits' => Integer,
+      'ConsumedCapacityUnits' => Float,
       'Count'                 => Integer,
       'Items'                 => [{
         'key'   => { 'S' => String },
